frontend/requirements.txt
flask
flask-login
kafka-python==2.0.2
grpcio-tools ~= 1.30
Jinja2 ~= 2.11
pytest ~= 5.4